implementation/instagram-feed/instagram-feed-10m/worker-service/requirements.txt
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
redis==5.0.1
kafka-python==2.0.2
